requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
APScheduler==3.6.3
async-timeout==4.0.2
attrs==21.4.0
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==2.0.12
configparser==5.2.0
dateparser==1.1.0
distlib==0.3.4
emoji==1.7.0
et-xmlfile==1.1.0
filelock==3.6.0
frozenlist==1.3.0
gunicorn==20.1.0
idna==3.3
multidict==6.0.2
numpy==1.22.3
openpyxl==3.0.9
pandas==1.4.1
platformdirs==2.5.1
pretty-html-table==0.9.15
python-binance==1.0.15
python-dateutil==2.8.2
python-telegram-bot==13.11
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
regex==2022.3.15
requests==2.27.1
six==1.16.0
tornado==6.1
tzdata==2021.5
tzlocal==4.1
ujson==5.1.0
urllib3==1.26.8
virtualenv==20.13.4
websockets==9.1
yarl==1.7.2
